+++ lib/src/channel/pusher-private-channel.dart
@@ -1,5 +1,5 @@
-import 'package:laravel_echo/src/channel/pusher-channel.dart';
-import 'package:laravel_echo/src/channel/private-channel.dart';
+import 'private-channel.dart';
+import 'pusher-channel.dart';
 
 ///
 /// This class represents a Pusher private channel.
@@ -10,7 +10,7 @@ class PusherPrivateChannel extends PusherChannel implements PrivateChannel {
 
   /// Trigger client event on the channel.
   PusherPrivateChannel whisper(String eventName, dynamic data) {
-    this.pusher.channels.channels[this.name].trigger('client-$eventName', data);
+    // this.subscription.channels[this.name].trigger('client-$eventName', data);
 
     return this;
   }
