+++ lib/src/channel/socketio-private-channel.dart
@@ -1,11 +1,12 @@
-import 'package:laravel_echo/src/channel/private-channel.dart';
-import 'package:laravel_echo/src/channel/socketio-channel.dart';
+import 'private-channel.dart';
+import 'socketio-channel.dart';
 
 ///
 /// This class represents a Socket.io presence channel.
 ///
 class SocketIoPrivateChannel extends SocketIoChannel implements PrivateChannel {
-  SocketIoPrivateChannel(dynamic socket, String name, dynamic options) : super(socket, name, options);
+  SocketIoPrivateChannel(dynamic socket, String name, dynamic options)
+      : super(socket, name, options);
 
   /// Trigger client event on the channel.
   SocketIoPrivateChannel whisper(String eventName, dynamic data) {
