+++ lib/src/connector/null-connector.dart
@@ -1,7 +1,7 @@
-import 'package:laravel_echo/src/connector/connector.dart';
-import 'package:laravel_echo/src/channel/null-channel.dart';
-import 'package:laravel_echo/src/channel/null-private-channel.dart';
-import 'package:laravel_echo/src/channel/null-presence-channel.dart';
+import '../channel/null-channel.dart';
+import '../channel/null-presence-channel.dart';
+import '../channel/null-private-channel.dart';
+import 'connector.dart';
 
 ///
 /// This class creates a null connector.
